Separate some components
