+++ CampusNavigator/CampusNavigator/CampusNavigator/Components/BottomNavItem.swift
@@ -1,8 +1,23 @@
-//
-//  BottomNavItem.swift
-//  CampusNavigator
-//
-//  Created by Sasindu 034 on 2025-02-23.
-//
+import SwiftUI
 
-import Foundation
+struct BottomNavItem: View {
+    let icon: String
+    let label: String
+    var isActive: Bool = false
+    
+    var body: some View {
+        Button {
+        } label: {
+            VStack(spacing: 4) {
+                Image(systemName: icon)
+                    .symbolVariant(isActive ? .fill : .none)
+                    .font(.system(size: 20))
+                
+                Text(label)
+                    .font(.caption2)
+            }
+            .foregroundColor(isActive ? .customDarkGreen : .secondary)
+            .frame(maxWidth: .infinity)
+        }
+    }
+}
